WIP: restructuring IEEE 754 functions: separation between constexpr/and non-const
