+++ include/universal/native/ieee754_msvc.hpp
@@ -21,26 +21,6 @@ namespace sw::universal {
 /////////////////////////////////////////////////////////////////////////////////////////////////////////
 // compiler specific long double IEEE floating point
 
-/*
-	Long double is not consistently implemented across different compilers.
-	The following section organizes the implementation details of each
-	of the compilers supported.
-
-	The x86 extended precision format is an 80-bit format first
-	implemented in the Intel 8087 math coprocessor and is supported
-	by all processors that are based on the x86 design that incorporate
-	a floating-point unit(FPU).This 80 - bit format uses one bit for
-	the sign of the significand, 15 bits for the exponent field
-	(i.e. the same range as the 128 - bit quadruple precision IEEE 754 format)
-	and 64 bits for the significand. The exponent field is biased by 16383,
-	meaning that 16383 has to be subtracted from the value in the
-	exponent field to compute the actual power of 2.
-	An exponent field value of 32767 (all fifteen bits 1) is reserved
-	so as to enable the representation of special states such as
-	infinity and Not a Number.If the exponent field is zero, the
-	value is a denormal number and the exponent of 2 is ¿16382.
-*/
-
 // Visual C++ does not support long double, it is just an alias for double
 /*
 union long_double_decoder {
@@ -54,22 +34,22 @@ union long_double_decoder {
 */
 
 // generate a binary string for a native long double precision IEEE floating point
-inline std::string to_hex(const long double& number) {
+inline std::string to_hex(long double number) {
 	return to_hex(double(number));
 }
 
 // generate a binary string for a native long double precision IEEE floating point
-inline std::string to_binary(const long double& number, bool bNibbleMarker = false) {
+inline std::string to_binary(long double number, bool bNibbleMarker = false) {
 	return to_binary(double(number), bNibbleMarker);
 }
 
 // return in triple form (+, scale, fraction)
-inline std::string to_triple(const long double& number) {
+inline std::string to_triple(long double number) {
 	return to_triple(double(number));
 }
 
 // generate a color coded binary string for a native double precision IEEE floating point
-inline std::string color_print(const long double& number) {
+inline std::string color_print(long double number) {
 	return color_print(double(number));
 }
